backend/requirements.txt
Django
djangorestframework
django-cors-headers
gunicorn
djoser
djangorestframework-simplejwt
django-filter
Pillow
